requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydub>=0.25.0
openai>=1.0.0
nltk>=3.6.0
python-multipart>=0.0.5
markitdown>=0.1.3
requests>=2.25.0
weasyprint>=52.5
python-docx>=0.8.11
beautifulsoup4>=4.9.0
markdown>=3.9
jinja2>=3.1.6
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.0
tinydb>=4.8.0
pydantic[email]>=2.0.0